requirements.txt
flask
gradio_client
pillow
requests
torch
torchvision
transformers
accelerate
